+++ src/test/extension.test.ts
@@ -314,68 +314,64 @@ suite("sortWords()", () => {
     });
 
     // Collate options
-    let tt: Array<[string, boolean, boolean, string, string]> = [
+    let tt: Array<[string, boolean, string, string]> = [
         ["delimiter: comma",
-            false, false, "1,2,10", "1,10,2"],
+            false, "1,2,10", "1,2,10"],
         ["delimiter: comma; keep consecutive occurrence",
-            false, false, "1,,2,10", ",1,10,2"],
+            false, "1,,2,10", ",1,2,10"],
         ["delimiter: comma; reproduce at most one trailing space",
-            false, false, "1,  2,10", "1, 10, 2"],
+            false, "1,  2,10", "1, 2, 10"],
         ["delimiter: tab",
-            false, false, "1\t2\t10", "1\t10\t2"],
+            false, "1\t2\t10", "1\t2\t10"],
         ["delimiter: tab; keep consecutive occurrence",
-            false, false, "1\t\t2\t10", "\t1\t10\t2"],
+            false, "1\t\t2\t10", "\t1\t2\t10"],
         ["delimiter: tab; ignore spaces after them",
-            false, false, "1\t   2\t10", "1\t10\t2"],
+            false, "1\t   2\t10", "1\t2\t10"],
         ["delimiter: pipe",
-            false, false, "1|2|10", "1|10|2"],
+            false, "1|2|10", "1|2|10"],
         ["delimiter: pipe; keep consecutive occurrence",
-            false, false, "1||2|10", "|1|10|2"],
+            false, "1||2|10", "|1|2|10"],
         ["delimiter: pipe; reproduce at most one preceding space",
-            false, false, "1  |2|10", "1 |10 |2"],
+            false, "1  |2|10", "1 |2 |10"],
         ["delimiter: pipe; reproduce at most one trailing space",
-            false, false, "1|  2|10", "1| 10| 2"],
+            false, "1|  2|10", "1| 2| 10"],
         ["delimiter: pipe; reproduce at most one preceding & trailing space",
-            false, false, "1  |  2|10", "1 | 10 | 2"],
+            false, "1  |  2|10", "1 | 2 | 10"],
         ["delimiter: space",
-            false, false, "1 2 10", "1 10 2"],
+            false, "1 2 10", "1 2 10"],
         ["delimiter: space; treat consecutive occurrence as single occurrence",
-            false, false, "1  2 10", "1 10 2"],
+            false, "1  2 10", "1 2 10"],
         ["delimiter: priority; ',\\t' --> ', '",
-            false, false, "1,\t2,10", "1, 10, 2"],
+            false, "1,\t2,10", "1, 2, 10"],
         ["delimiter: priority; ',|' --> ','",
-            false, false, "1,|2,10", "|2,1,10"],
+            false, "1,|2,10", "|2,1,10"],
         ["delimiter: priority; ', ' --> ', '",
-            false, false, "1, 2,10", "1, 10, 2"],
+            false, "1, 2,10", "1, 2, 10"],
         ["delimiter: priority; '\\t,' --> '\\t'",
-            false, false, "1\t,2\t10", ",2\t1\t10"],
+            false, "1\t,2\t10", ",2\t1\t10"],
         ["delimiter: priority; '\\t|' --> '\\t'",
-            false, false, "1\t|2\t10", "|2\t1\t10"],
+            false, "1\t|2\t10", "|2\t1\t10"],
         ["delimiter: priority; '\\t ' --> '\\t'",
-            false, false, "1\t 2\t10", "1\t10\t2"],
+            false, "1\t 2\t10", "1\t2\t10"],
         ["delimiter: priority; '|,' --> '|'",
-            false, false, "1|,2|10", ",2|1|10"],
+            false, "1|,2|10", ",2|1|10"],
         ["delimiter: priority; '|\\t' --> '| '",
-            false, false, "1|\t2|10", "1| 10| 2"],
+            false, "1|\t2|10", "1| 2| 10"],
         ["delimiter: priority; '| ' --> '| '",
-            false, false, "1| 2|10", "1| 10| 2"],
+            false, "1| 2|10", "1| 2| 10"],
         ["delimiter: priority; ' ,' --> ','",
-            false, false, "1 ,2 10", "1,2 10"],
+            false, "1 ,2 10", "1,2 10"],
         ["delimiter: priority; ' \\t' --> ' '",
-            false, false, "1 \t2 10", "1 10 2"],
+            false, "1 \t2 10", "1 2 10"],
         ["delimiter: priority; ' |' --> ' |'",
-            false, false, "1 |2|10", "1 |10 |2"],
-        ["options: asc",
-            false, false, "2,10,あ,ab,２,Ac,ア", "10,2,２,ab,Ac,あ,ア"],
-        ["options: asc, numeric",
-            false, true, "2,10,あ,ab,２,Ac,ア", "2,２,10,ab,Ac,あ,ア"],
-        ["options: desc",
-            true, false, "2,10,あ,ab,２,Ac,ア", "あ,ア,Ac,ab,2,２,10"],
-        ["options: desc, numeric",
-            true, true, "2,10,あ,ab,２,Ac,ア", "あ,ア,Ac,ab,10,2,２"],
+            false, "1 |2|10", "1 |2 |10"],
+        ["options: ascending",
+            false, "2,10,あ,ab,２,Ac,ア", "10,2,２,ab,Ac,あ,ア"],
+        ["options: descending",
+            true, "2,10,あ,ab,２,Ac,ア", "あ,ア,Ac,ab,2,２,10"],
     ];
     tt.forEach(t => {
-        const [title, descending, numeric, input, expected] = t;
+        const [title, descending, input, expected] = t;
         test(title, async () => {
             // Set the test input text and the selection
             const editor = vscode.window.activeTextEditor!;
@@ -392,7 +388,7 @@ suite("sortWords()", () => {
             editor.selections = [new Selection(0, 0, eod.line, eod.character)];
 
             // Call the logic
-            await my.sortWords(editor, descending, numeric);
+            await my.sortWords(editor, descending);
 
             const result = {
                 text: editor.document.getText(),